Rename `deploy` job to better express intent
This job's responsibility is all about building and publishing a
container image. Deployment would be a job following this, but can't be
part of the template because of the many options of how to deploy and
run an image.

Also note that the job also provides the produced container image as an
output, so depending jobs can make use of it (f. ex. some sort of
reusable workflow that needs to reference an image by its tag).