Keep vulnerability scan attestion up-to-date
As soon as the vulnerability scan passes in the periodic scan we want to
keep our attestation up-to-date in the container registry. This is
useful for when implementing a policy that will look at the age of a
vulnerability scan to decide whether a workload will be allowed to run
or not.